requirements.txt
pymongo
python-dotenv
jupyterlab
ipykernel
pandas
PyYAML
pymongo[srv]
numpy
scikit-learn
-e .